feat: 첨부파일 DAO & DTO 수정

+++ shop29/src/main/java/com/toy2/shop29/qna/domain/AttachmentDto.java
@@ -16,7 +16,7 @@ public class AttachmentDto {
     private String filePath;
     private Integer width;
     private Integer height;
-    private Integer size;
+    private long size;
     private String extension;
     private Boolean isActive;
